判断删除参数不能为空

+++ zoe-phip/2.module/phip-module-service-impl/src/main/java/com/zoe/phip/module/service/impl/in/BaseInServiceImpl.java
@@ -65,6 +65,8 @@ public abstract class BaseInServiceImpl<T extends MasterEntity, TMapper extends 
 
     @Override
     public int deleteById(String id) throws Exception {
+        if(StringUtil.isNullOrWhiteSpace(id))
+            throw new BusinessException("删除参数不能为空");
         return mapper.deleteByPrimaryKey(id);
     }
